Revert "refactor: download_file function + use of env variables for configuration (data folders, proxy url, ...) + integration of dodgr"
This reverts commit 5205b6a62ed499941956110a02b71ec0420d2ae1.
