+++ mobility/transport_zones.py
@@ -1,8 +1,14 @@
+import os
+import pathlib
 import pandas as pd
+import geopandas as gpd
 import shapely
+import numpy as np
 
-from mobility.parsers.admin_boundaries import get_french_cities_boundaries, get_french_epci_boundaries
-from mobility.parsers.urban_units import get_french_urban_units
+from mobility.parsers.ign import prepare_ign
+from mobility.parsers.urban_units import prepare_urban_units
+from mobility.parsers.get_osm import get_osm
+
     
 def get_transport_zones(insee_city_id, method="epci_rings", radius=40):
     """
@@ -30,11 +36,28 @@ def get_transport_zones(insee_city_id, method="epci_rings", radius=40):
 
     """
     
+    data_folder_path = (
+        pathlib.Path(os.path.dirname(__file__)).parents[0] / "mobility/data/"
+    )
+    
+    # Check if the parquet files already exist, if not writes them calling the corresponding funtion
+    check_files = (data_folder_path / "ign/admin-express/EPCI.parquet").exists()
+    check_files = (data_folder_path / "ign/admin-express/COMMUNE.parquet").exists()
+    
+    if not (check_files):
+        prepare_ign()
 
+    
     # Load IGN city admin borders (admin express)
-    cities = get_french_cities_boundaries()
+    cities = gpd.read_parquet(
+        data_folder_path / "ign/admin-express/COMMUNE.parquet",
+        columns=["INSEE_COM", "SIREN_EPCI", "NOM", "geometry"]
+    )
     
     if method == "epci_rings":
+    
+        # Fix the Grand Paris EPCI ids
+        cities.loc[cities["SIREN_EPCI"].str[0:9] == "200054781", "SIREN_EPCI"] = "200054781"
         
         # Find the EPCI of the city
         epci_id = cities.loc[cities["INSEE_COM"] == insee_city_id, "SIREN_EPCI"].values
@@ -45,7 +68,7 @@ def get_transport_zones(insee_city_id, method="epci_rings", radius=40):
             epci_id = epci_id[0]
             
         # Load the geometries of all nearby EPCIs
-        epcis = get_french_epci_boundaries()
+        epcis = gpd.read_parquet(data_folder_path / "ign/admin-express/EPCI.parquet")
         
         # Select first / second ring of EPCIs around the EPCI of the city
         city_epci = epcis.loc[epcis["CODE_SIREN"] == epci_id].iloc[0]
@@ -76,8 +99,25 @@ def get_transport_zones(insee_city_id, method="epci_rings", radius=40):
         )
 
     # Add the urban unit category info
-    urban_units = get_french_urban_units()
-
+    check_files = (data_folder_path / "insee/territories/UU2020_au_01-01-2023.xlsx").exists()
+    
+    if not (check_files):
+        prepare_urban_units()
+        
+    urban_units = pd.read_excel(
+        data_folder_path / "insee/territories/UU2020_au_01-01-2023.xlsx",
+        sheet_name="Composition_communale",
+        skiprows=5
+    )
+    
+    urban_units = urban_units.iloc[:, [0, 5]]
+    urban_units.columns = ["INSEE_COM", "urban_unit_category"]
+    urban_units["urban_unit_category"] = np.where(
+        urban_units["urban_unit_category"] != "H",
+        urban_units["urban_unit_category"],
+        "R"
+    )
+    
     cities = pd.merge(cities, urban_units, on="INSEE_COM", how="left")
     
     # Prepare and format the transport zones data frame
@@ -91,4 +131,22 @@ def get_transport_zones(insee_city_id, method="epci_rings", radius=40):
     ]]
     
     return transport_zones
+    
 
+# Monkey patch openpyxl to avoid an error when opening the urban units INSEE file
+# Source : https://stackoverflow.com/questions/71733414/copying-from-a-range-of-cells-with-openpyxl-error-colors-must-be-argb-hex-valu
+from openpyxl.styles.colors import WHITE, RGB
+__old_rgb_set__ = RGB.__set__
+def __rgb_set_fixed__(self, instance, value):
+    try:
+        __old_rgb_set__(self, instance, value)
+    except ValueError as e:
+        if e.args[0] == 'Colors must be aRGB hex values':
+            __old_rgb_set__(self, instance, WHITE)
+RGB.__set__ = __rgb_set_fixed__
+
+
+# Small test
+z = get_transport_zones("75056", method="radius", radius=40)
+transport_zones_boundary = z.to_crs(4326).unary_union
+osm_file_path = get_osm(transport_zones_boundary, verify="C:/Users/pouchaif/Documents/dev/forcepoint.pem")
